requirements.txt
# Core framework
google-adk>=1.0.0
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.18.0
google-genai>=1.0.0

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0

# HTTP client
httpx>=0.27.0

# X (Twitter) API
tweepy>=4.14.0

# Telegram Bot
python-telegram-bot>=20.0

# Environment
python-dotenv>=1.0.0

# Data validation
pydantic>=2.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
